KB-12295 Refactor and shifted new retire content and versioning to new extended class. (#151) (#152)
* KB-12295 implemented for storing the access setting data for new version course

* KB-12295 implemented for storing the retirement logs in audit table.

* KB-12295 implemented for storing the retirement logs in audit table currecting formating.

* KB-12295 fixed bug for copying the access setting data.

* KB-12295 Refactor and shifted new retire content and versioning to new extended class.

* KB-12295 Refactor and shifted new retire content and versioning to new extended class.

---------
