+++ content-api/content-service/app/modules/ContentModule.scala
@@ -3,7 +3,7 @@ package modules
 import com.google.inject.AbstractModule
 import org.sunbird.channel.actors.ChannelActor
 import org.sunbird.collectioncsv.actors.CollectionCSVActor
-import org.sunbird.content.actors.{AppActor, AssetActor, CategoryActor, CollectionActor, ContentActor, EventActor, EventSetActor, HealthActor, LicenseActor, ObjectActor}
+import org.sunbird.content.actors.{AppActor, AssetActor, CategoryActor, CollectionActor, ContentActor, EventActor, EventSetActor, ExtendedContentActor, HealthActor, LicenseActor, ObjectActor}
 import play.libs.akka.AkkaGuiceSupport
 import utils.ActorNames
 
@@ -24,6 +24,7 @@ class ContentModule extends AbstractModule with AkkaGuiceSupport {
         bindActor(classOf[AppActor], ActorNames.APP_ACTOR)
         bindActor(classOf[ObjectActor], ActorNames.OBJECT_ACTOR)
         bindActor(classOf[CollectionCSVActor], ActorNames.COLLECTION_CSV_ACTOR)
+        bindActor(classOf[ExtendedContentActor], ActorNames.EXTENDED_CONTENT_ACTOR)
         println("Initialized application actors...")
         // $COVERAGE-ON
     }
